moved test data and assertion functinos to separate files
